주석 변경 & User 도메인 JoinUserUseCase refactoring
