+++ app-common/src/main/java/kr/jay/appcommon/model/user/OAuthProvider.java
@@ -5,7 +5,7 @@ package kr.jay.appcommon.model.user;
  *
  * @author jaypark
  * @version 1.0.0
- * @date 2023/06/20
+ * @since 2023/06/20
  */
 public enum OAuthProvider {
 	GOOGLE,
