+++ app-core/src/main/java/kr/jay/appcore/domain/user/entity/User.java
@@ -1,4 +1,4 @@
-package kr.jay.appcore.domain.user;
+package kr.jay.appcore.domain.user.entity;
 
 import java.time.LocalDateTime;
 
@@ -9,7 +9,7 @@ import kr.jay.appcommon.model.user.OAuthProvider;
  *
  * @author jaypark
  * @version 1.0.0
- * @date 2023/06/20
+ * @since 2023/06/20
  */
 public record User(
 	Long id,
@@ -21,13 +21,16 @@ public record User(
 	LocalDateTime lastLoginAt
 ) {
 
+	private static final int LENGTH_NICK_NAME_MIN = 2;
+	private static final int LENGTH_NICK_NAME_MAX = 64;
+
 	public static User create(
 		final String providerId,
 		final String email,
 		final String picture,
 		final OAuthProvider provider,
 		final String nickName
-	){
+	) {
 		return new User(
 			null,
 			providerId,
