+++ app-core/src/main/java/kr/jay/appcore/domain/user/exception/SameNicknameUserAlreadyExistException.java
@@ -5,7 +5,7 @@ package kr.jay.appcore.domain.user.exception;
  *
  * @author jaypark
  * @version 1.0.0
- * @date 2023/06/20
+ * @since 2023/06/20
  */
 public class SameNicknameUserAlreadyExistException extends RuntimeException	{
 	private static final String MESSAGE_FORMAT = "이미 사용중인 NickName입니다. nickName: %s";
