+++ app-core/src/main/java/kr/jay/appcore/domain/user/infrastructure/IdTokenVerifier.java
@@ -1,5 +1,7 @@
 package kr.jay.appcore.domain.user.infrastructure;
 
+import java.util.Optional;
+
 import kr.jay.appcommon.model.user.OAuthProvider;
 
 /**
@@ -7,11 +9,11 @@ import kr.jay.appcommon.model.user.OAuthProvider;
  *
  * @author jaypark
  * @version 1.0.0
- * @date 2023/06/20
+ * @since 2023/06/20
  */
 public interface IdTokenVerifier {
 
-	OAuthInfo verify( final OAuthProvider oAuthProvider, final String idToken);
+	Optional<OAuthInfo> verify( final OAuthProvider oAuthProvider, final String idToken);
 
 	record OAuthInfo(
 		String providerId,
