+++ app-core/src/main/java/kr/jay/appcore/domain/user/usecase/FindUserUseCase.java
@@ -1,13 +1,13 @@
 package kr.jay.appcore.domain.user.usecase;
 
-import kr.jay.appcore.domain.user.User;
+import kr.jay.appcore.domain.user.entity.User;
 
 /**
  * FindUserUseCase
  *
  * @author jaypark
  * @version 1.0.0
- * @date 2023/06/20
+ * @since 2023/06/20
  */
 public interface FindUserUseCase {
 	User query(final Long userId);
