+++ app-core/src/main/java/kr/jay/appcore/domain/user/usecase/JoinUserUseCase.java
@@ -7,7 +7,7 @@ import kr.jay.appcommon.model.user.OAuthProvider;
  *
  * @author jaypark
  * @version 1.0.0
- * @date 2023/06/20
+ * @since 2023/06/20
  */
 public interface JoinUserUseCase {
